# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PotholeSeverity = typeof PotholeSeverity[keyof typeof PotholeSeverity];


export const PotholeSeverity = {
  minor: 'minor',
  moderate: 'moderate',
  severe: 'severe',
} as const;

export type PotholeStatus = typeof PotholeStatus[keyof typeof PotholeStatus];


export const PotholeStatus = {
  reported: 'reported',
  confirmed: 'confirmed',
  'in-progress': 'in-progress',
  fixed: 'fixed',
} as const;

export interface Pothole {
  id: string;
  lat: number;
  lng: number;
  severity: PotholeSeverity;
  status: PotholeStatus;
  confirmations: number;
  createdAt: string;
  streetName: string;
  neighborhood: string;
  notes?: string;
}

export type CreatePotholeRequestSeverity = typeof CreatePotholeRequestSeverity[keyof typeof CreatePotholeRequestSeverity];


export const CreatePotholeRequestSeverity = {
  minor: 'minor',
  moderate: 'moderate',
  severe: 'severe',
} as const;

export interface CreatePotholeRequest {
  lat: number;
  lng: number;
  severity: CreatePotholeRequestSeverity;
  notes?: string;
  streetName: string;
  neighborhood: string;
}

export interface User {
  id: string;
  email: string;
  displayName: string;
  createdAt: string;
}

/**
 * Wrapper so that "nobody is signed in" is a valid 200, not a 401.
 */
export interface CurrentUser {
  user: User | null;
}

export interface SignUpRequest {
  /**
     * @maxLength 254
     * @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$
     */
  email: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  password: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName: string;
}

export interface LogInRequest {
  email: string;
  password: string;
}

export interface ErrorResponse {
  error: string;
}