# lib/api-zod/src/generated/types/currentUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

/**
 * Wrapper so that "nobody is signed in" is a valid 200, not a 401.
 */
export interface CurrentUser {
  user: User | null;
}