# lib/api-zod/src/generated/types/signUpRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignUpRequest {
  /**
     * @maxLength 254
     * @pattern ^[^@\s]+@[^@\s]+\.[^@\s]+$
     */
  email: string;
  /**
     * @minLength 8
     * @maxLength 200
     */
  password: string;
  /**
     * @minLength 1
     * @maxLength 60
     */
  displayName: string;
}